Move model classes into package
